requirements.txt
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.0,<3.1
Flask-Migrate>=4.0,<4.1
Flask-WTF>=1.1,<1.2
WTForms>=3.0,<3.1
python-dotenv>=1.0,<2.0
python-dateutil>=2.8,<3.0
PyYAML>=6.0,<7.0
pytest>=7.4,<8.0
gunicorn>=21.0,<22.0
